firestationsEndpoints

+++ src/main/java/com/safetynetalerts/api/data/dao/PersonDao.java
@@ -7,11 +7,12 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 import java.util.Optional;
 
+//TODO rename repo
 @Repository
 public interface PersonDao extends CrudRepository<PersonEntity, Long> {
-    List<PersonEntity> findAllByFireStation(int fireStation);
+    List<PersonEntity> findAllByFireStation(Long fireStation);
 
-    List<PersonEntity> findAllByFireStationIn(List<Integer> stationNumbers);
+    List<PersonEntity> findAllByFireStationIn(List<Long> stationNumbers);
 
     List<PersonEntity> findAllByAddress(String Address);
